Enhance crashlytics logging
