+++ AirCasting/Bootstrap/AppBootstrap.swift
@@ -5,6 +5,7 @@ import Foundation
 import Resolver
 import GoogleMaps
 import GooglePlaces
+import FirebaseCrashlytics
 
 class AppBootstrap {
     @Injected private var firstRunInfoProvider: FirstRunInfoProvidable
@@ -15,6 +16,11 @@ class AppBootstrap {
         if firstRunInfoProvider.isFirstAppLaunch {
             handleFirstAppLaunch()
         }
+        #if BETA
+        Crashlytics.crashlytics().setCustomValue("BETA", forKey: "build_type")
+        #elseif RELEASE
+        Crashlytics.crashlytics().setCustomValue("RELEASE", forKey: "build_type")
+        #endif
         firstRunInfoProvider.registerAppLaunch()
         GMSServices.provideAPIKey(GOOGLE_MAP_KEY)
         GMSPlacesClient.provideAPIKey(GOOGLE_PLACES_KEY)
